backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]>=3.1.0
gunicorn==20.1.0